+++ ElectrodZMultiplayer/Core/Data/Messages/RestartGameMessageData.cs
@@ -16,14 +16,14 @@ namespace ElectrodZMultiplayer.Data.Messages
         /// Time to start game in seconds
         /// </summary>
         [JsonProperty("time")]
-        public float Time { get; set; }
+        public double Time { get; set; }
 
         /// <summary>
         /// Is valid
         /// </summary>
         public override bool IsValid =>
             base.IsValid &&
-            (Time >= 0.0f);
+            (Time >= 0.0);
 
         /// <summary>
         /// Constructs a restart game request message for deserializers
@@ -37,9 +37,9 @@ namespace ElectrodZMultiplayer.Data.Messages
         /// Constructs a restart game request message
         /// </summary>
         /// <param name="time">Time to start game in seconds</param>
-        public RestartGameMessageData(float time) : base(Naming.GetMessageTypeNameFromMessageDataType<RestartGameMessageData>())
+        public RestartGameMessageData(double time) : base(Naming.GetMessageTypeNameFromMessageDataType<RestartGameMessageData>())
         {
-            if (time < 0.0f)
+            if (time < 0.0)
             {
                 throw new ArgumentException("Time must be positive", nameof(time));
             }
