+++ ElectrodZMultiplayer/Core/Data/Messages/StartGameMessageData.cs
@@ -16,14 +16,14 @@ namespace ElectrodZMultiplayer.Data.Messages
         /// Time to start game in seconds
         /// </summary>
         [JsonProperty("time")]
-        public float Time { get; set; }
+        public double Time { get; set; }
 
         /// <summary>
         /// Is valid
         /// </summary>
         public override bool IsValid =>
             base.IsValid &&
-            (Time >= 0.0f);
+            (Time >= 0.0);
 
         /// <summary>
         /// Constructs a start game request message for deserializers
@@ -37,9 +37,9 @@ namespace ElectrodZMultiplayer.Data.Messages
         /// Constructs a start game request message
         /// </summary>
         /// <param name="time">Time to start game in seconds</param>
-        public StartGameMessageData(float time) : base(Naming.GetMessageTypeNameFromMessageDataType<StartGameMessageData>())
+        public StartGameMessageData(double time) : base(Naming.GetMessageTypeNameFromMessageDataType<StartGameMessageData>())
         {
-            if (time < 0.0f)
+            if (time < 0.0)
             {
                 throw new ArgumentException("Time must be positive", nameof(time));
             }
