+++ ElectrodZMultiplayer/Core/Delegates/GameStartRequestedDelegate.cs
@@ -7,5 +7,5 @@ namespace ElectrodZMultiplayer
     /// Used to signal a game start request
     /// </summary>
     /// <param name="time">Time to start game in seconds</param>
-    public delegate void GameStartRequestedDelegate(float time);
+    public delegate void GameStartRequestedDelegate(double time);
 }
