+++ ElectrodZMultiplayer/Core/Delegates/GameStopRequestedDelegate.cs
@@ -7,5 +7,5 @@ namespace ElectrodZMultiplayer
     /// Used to signal a game stop request
     /// </summary>
     /// <param name="time">Time to stop game in seconds</param>
-    public delegate void GameStopRequestedDelegate(float time);
+    public delegate void GameStopRequestedDelegate(double time);
 }
